requirements.txt
absl-py==0.11.0
astor==0.8.1
backcall==0.2.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
conda==4.9.2
conda-package-handling==1.7.0
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.22
decorator==4.4.2
gast==0.2.2
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
h5py==2.10.0
httplib2==0.19.0
idna==2.9
importlib-metadata==3.7.2
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.17.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.2
Markdown==3.3.4
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.5
numpy==1.18.5
oauth2client==4.1.3
olefile==0.46
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.8
protobuf==3.15.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.20
Pygments==2.8.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
regex==2020.11.13
requests==2.23.0
rsa==4.7.2
ruamel-yaml==0.15.87
scipy==1.6.1
six==1.14.0
tensorboard==1.15.0
tensorflow==1.15.5
tensorflow-estimator==1.15.1
termcolor==1.1.0
tokenizers==0.10.2
toml==0.10.2
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1